+++ feature/main/src/main/java/com/napzak/market/main/MainScreen.kt
@@ -61,8 +61,6 @@ import com.napzak.market.store.edit_store.navigation.editStoreGraph
 import com.napzak.market.store.edit_store.navigation.navigateToEditStore
 import com.napzak.market.store.store.navigation.navigateToStore
 import com.napzak.market.store.store.navigation.storeGraph
-import com.napzak.market.wishlist.navigation.navigateToWishlist
-import com.napzak.market.wishlist.navigation.wishlistGraph
 import kotlinx.collections.immutable.toImmutableList
 
 @Composable
@@ -315,14 +313,8 @@ private fun MainNavHost(
             navigateToSales = { /* TODO: 판매내역 화면으로 이동 */ },
             navigateToPurchase = { /* TODO: 구매내역 화면으로 이동 */ },
             navigateToRecent = { /* TODO: 최근 본 상품 화면으로 이동 */ },
-            navigateToWishlist = navigator.navController::navigateToWishlist,
-            restartApplication = restartApplication,
-            modifier = modifier,
-        )
-
-        wishlistGraph(
-            navigateToUp = navigator::navigateUp,
             navigateToProductDetail = navigator.navController::navigateToProductDetail,
+            restartApplication = restartApplication,
             modifier = modifier,
         )
