+++ feature/mypage/src/main/java/com/napzak/market/mypage/wishlist/WishlistSideEffect.kt
@@ -1,4 +1,4 @@
-package com.napzak.market.wishlist
+package com.napzak.market.mypage.wishlist
 
 sealed interface WishlistSideEffect {
     data object ShowHeartToast : WishlistSideEffect
