+++ feature/mypage/src/main/java/com/napzak/market/mypage/wishlist/component/WishlistTopSection.kt
@@ -1,4 +1,4 @@
-package com.napzak.market.wishlist.component
+package com.napzak.market.mypage.wishlist.component
 
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.Spacer
@@ -10,7 +10,6 @@ import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.graphics.vector.ImageVector
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.res.vectorResource
@@ -19,8 +18,9 @@ import com.napzak.market.common.type.TradeType
 import com.napzak.market.designsystem.R.drawable.ic_left_chevron
 import com.napzak.market.designsystem.component.tabbar.TradeTypeTabBar
 import com.napzak.market.designsystem.theme.NapzakMarketTheme
-import com.napzak.market.feature.wishlist.R.string.wishlist_back_button
-import com.napzak.market.feature.wishlist.R.string.wishlist_title
+import com.napzak.market.designsystem.theme.shadowBlack
+import com.napzak.market.feature.mypage.R.string.wishlist_back_button
+import com.napzak.market.feature.mypage.R.string.wishlist_title
 import com.napzak.market.ui_util.ShadowDirection
 import com.napzak.market.ui_util.napzakGradientShadow
 import com.napzak.market.ui_util.noRippleClickable
@@ -37,8 +37,8 @@ internal fun WishlistTopSection(
             .fillMaxWidth()
             .napzakGradientShadow(
                 height = 2.dp,
-                startColor = Color(0xFF000000), // TODO: shadow 컬러값 변경
-                endColor = Color.Transparent,
+                startColor = shadowBlack,
+                endColor = NapzakMarketTheme.colors.transWhite,
                 direction = ShadowDirection.Bottom,
             )
             .padding(start = 20.dp, bottom = 18.dp),
@@ -68,8 +68,8 @@ internal fun WishlistTopSection(
             .fillMaxWidth()
             .napzakGradientShadow(
                 height = 2.dp,
-                startColor = Color(0xFF000000), // TODO: shadow 컬러값 변경
-                endColor = Color.Transparent,
+                startColor = shadowBlack,
+                endColor = NapzakMarketTheme.colors.transWhite,
                 direction = ShadowDirection.Bottom,
             )
             .padding(horizontal = 20.dp),
