+++ settings.gradle.kts
@@ -76,5 +76,4 @@ include(
     ":feature:splash",
     ":feature:login",
     ":feature:chat",
-    ":feature:wishlist",
 )
